requirements.txt
selenium==3.12.0
Appium-Python-Client==0.26
pandas==0.22.0
pymssql==2.1.4.dev5
PyMySQL==0.8.0
PyYAML==3.12
openpyxl==2.5.3